Update spec and include 'data_devices' for osd nodes
When an osd node is described, the data_devices section
must be specified as per drivegroups doc. This change
introduces the j2 section (which is WIP right now) and
if data_devices is empty --all-available-devices option
is used by orchestrator.
The next step is to add more logic and support many other
filters.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>