Add zap tasks on tripleo_cluster_osds role
This change just adds a new set of tasks to
the tripleo_cluster_osds role which automates
the osds zap action.
This is really useful for some environments
that come up w/ a not clean disk map, resulting
in a locked device tree which prevents osds to
be added in the ceph cluster.

Closes #1

Signed-off-by: Francesco Pantano <fpantano@redhat.com>