Extend spec definition improving the related data structures
The purpose of this commit is to improve the way the spec is built for
multiple Ceph services.
In particular, 'service_type' and 'service_id' are added to cover Ceph
services != (mon, mgr, osd).
The data structure 'tripleo_cephadm_services' is now extended and the
spec template is able to process a new 'spec' structure associated to
a given service.
By doing this an arbitrary number of options can be added to a specific
service and passed as input.
The count option is now available and this value can be retrieved from
the structure (the operator can define the count number during the
deployment phase).
If no count is specified, the value is set to the number of hosts in
the related group.
For each osd data structure ({data,db,wal}_devices) there's a filter
map associated and can be rendered via the proper function.
The defined macro is also generic, and it's reused to render a generic
'section' (e.g., spec) which can be associated to a given Ceph service.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>