Fix count option indentation and refact site.yaml
The count option indentation was wrong and this
change just fixed it.
The main site.yaml playbook now includes in the
main deployment flow the spec, which can be
applied according to the data structures passed
as input.
Then, the flow is resumed with pool creation
and some additional tasks.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>